tweak: added more docs to code, created internal Functions class
